Added calculations of song difficulty and notes/chords per phrase/iteration.  Also added a fix for the no-notes charts

+++ RocksmithToolkitLib/DLCPackage/Manifest/Attributes.cs
@@ -79,7 +79,7 @@ namespace RocksmithToolkitLib.DLCPackage.Manifest
         public bool Vibrato { get; set; }
         public string VocalsAssetId { get; set; }
         public bool FirstArrangementInSong { get; set; }
-        public int SongDifficulty { get; set; }
+        public float SongDifficulty { get; set; }
         public int AverageTempo { get; set; }
     }
 }
